# document/openapi/openapi-auth.yaml
openapi: 3.0.3
info:
  title: Gather Auth API
  description: |
    Gather 서비스 인증 및 회원가입 관련 API 명세입니다.

    프론트 연동 참고사항:
    - `signup`, `login`, `logout`, `reissue`는 인증 행위 API이므로 동사형 경로를 사용합니다.
    - `regions`, `categories` 조회 API는 auth 도메인 소유는 아니지만 회원가입 화면에서 필요하므로 함께 명세에 포함했습니다.
    - 회원가입 요청에는 관심 카테고리 enum 값 배열인 `interestCategories`를 전달합니다. 값은 `PostingCategory`(ENVIRONMENT, EDUCATION, CULTURE, COMMUNITY, WELFARE, OVERSEAS) 중 중복 없이 1개 이상이며, 정의되지 않은 값은 400(`VALIDATION_ERROR`)입니다.
    - `GET /api/v1/categories`는 아직 기존 16종 분류(`id`/`code`/`name`)를 반환합니다. 6종 enum 전환은 별도 작업으로 진행 중이므로, 전환 완료 전까지 회원가입 카테고리 선택지는 프론트에서 6종 enum 기준으로 처리합니다.
    - Refresh Token은 응답 body가 아니라 `gather_refresh_token` HttpOnly 쿠키로 전달합니다.
    - 재발급과 로그아웃은 Access Token 없이 Refresh Token 쿠키만으로 요청합니다.
    - 로그아웃은 같은 Refresh Token 쿠키로 여러 번 요청해도 성공 응답을 반환합니다.
    - 휴대폰 인증은 Gather가 5분 유효한 인증문구를 만들고, 사용자가 OCTOMO 대표번호로 직접 문자를 보낸 뒤 confirm API로 확인하는 번호 점유 인증입니다. 인증 완료 결과는 30분 안에 해당 `verificationId`와 전화번호를 함께 제출한 일반·카카오 회원가입에서 한 번만 소비할 수 있습니다.
    - offset 없는 이메일 인증 시각(`expiresAt`, `resendAvailableAt`, `verifiedAt`)은 UTC 기준입니다. 휴대폰 인증 `expiresAt`은 UTC offset(`Z`)을 포함합니다.
    - 카카오 로그인은 `POST /api/v1/auth/kakao/login`에 `{authorizationCode, redirectUri}`를 전달합니다. `redirectUri`는 서버 허용 목록과 문자열까지 정확히 일치해야 합니다. 응답은 200이며 `signupStatus`로 기존 회원(`LOGIN_COMPLETED`)과 신규 회원(`ADDITIONAL_INFO_REQUIRED`)을 구분합니다. `ADDITIONAL_INFO_REQUIRED`는 에러가 아닙니다.
...
  version: 1.0.0
servers:
  - url: /
    description: Relative server URL
...
paths:
...
  /api/v1/auth/signup:
    post:
      tags:
        - Auth
      summary: 회원가입
      description: 요청 전화번호와 일치하는 30분 이내 미소비 휴대폰 인증이 필요합니다. 기본 정보, 계정 정보, 프로필 정보, 활동 지역, 관심 카테고리, 약관 동의를 한 번에 제출하여 회원가입과 자동 로그인을 완료합니다. Access Token은 응답 body로, Refresh Token은 HttpOnly 쿠키로 발급됩니다.
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
            example:
              name: 홍길동
              birthDate: '2002-03-15'
              gender: MALE
              phoneNumber: '01012345678'
              phoneVerificationId: 5c5d5db1-4187-43d0-8580-672307994878
              email: test@example.com
              password: password123!
              passwordConfirm: password123!
              nickname: 길동
              introduction: 함께 봉사하고 싶습니다.
              activityRegionId: 123
              interestCategories:
                - WELFARE
                - EDUCATION
              serviceTermsAgreed: true
              privacyPolicyAgreed: true
              marketingAgreed: false
      responses:
        '201':
          description: 회원가입 및 자동 로그인 성공
          headers:
            Set-Cookie:
              description: HttpOnly Refresh Token 쿠키. Domain은 설정하지 않으며 Path는 `/api/v1/auth`입니다.
              schema:
                type: string
              example: gather_refresh_token=refresh-token-value; Path=/api/v1/auth; Max-Age=1209600; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupApiResponse'
              example:
                success: true
                data:
                  userId: 1
                  email: test@example.com
                  name: 홍길동
                  nickname: 길동
                  accessToken: access-token-value
                  tokenType: Bearer
                error: null
        '400':
          description: 요청 값 검증 실패 또는 비즈니스 규칙 위반
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
              examples:
                validationError:
                  summary: 단일 필드 검증 실패
                  value:
                    success: false
                    data: null
                    error:
                      code: VALIDATION_ERROR
                      message: 요청 값이 올바르지 않습니다.
                passwordMismatch:
                  summary: 비밀번호 확인 불일치
                  value:
                    success: false
                    data: null
                    error:
                      code: PASSWORD_MISMATCH
                      message: 비밀번호가 일치하지 않습니다.
                emailNotVerified:
                  summary: 이메일 인증 미완료
                  value:
                    success: false
                    data: null
                    error:
                      code: EMAIL_NOT_VERIFIED
                      message: 이메일 인증이 완료되지 않았습니다.
                phoneVerificationRequired:
                  summary: 휴대폰 인증 없음·ID/전화번호 불일치·인증 만료·이미 소비됨
                  value:
                    success: false
                    data: null
                    error:
                      code: PHONE_VERIFICATION_REQUIRED
                      message: 휴대폰 인증이 필요합니다.
                requiredTermsNotAgreed:
                  summary: 필수 약관 미동의
                  value:
                    success: false
                    data: null
                    error:
                      code: REQUIRED_TERMS_NOT_AGREED
                      message: 필수 약관 동의가 필요합니다.
                invalidInterestCategoryCount:
                  summary: 관심 카테고리 개수 오류
                  value:
                    success: false
                    data: null
                    error:
                      code: INVALID_INTEREST_CATEGORY_COUNT
                      message: 관심 카테고리는 중복 없이 1개 이상 선택해야 합니다.
        '404':
          description: 활동 지역 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
              examples:
                regionNotFound:
                  summary: 활동 지역 없음
                  value:
                    success: false
                    data: null
                    error:
                      code: REGION_NOT_FOUND
                      message: 활동 지역을 찾을 수 없습니다.
        '409':
          description: 중복 데이터 존재
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
              examples:
                duplicateEmail:
                  summary: 이메일 중복
                  value:
                    success: false
                    data: null
                    error:
                      code: DUPLICATE_EMAIL
                      message: 이미 사용 중인 이메일입니다.
                duplicatePhoneNumber:
                  summary: 전화번호 중복
                  value:
                    success: false
                    data: null
                    error:
                      code: DUPLICATE_PHONE_NUMBER
                      message: 이미 사용 중인 전화번호입니다.
                duplicateNickname:
                  summary: 닉네임 중복
                  value:
                    success: false
                    data: null
                    error:
                      code: DUPLICATE_NICKNAME
                      message: 이미 사용 중인 닉네임입니다.
                accountRejoinBlocked:
                  summary: 탈퇴 후 7일 재가입 제한
                  value:
                    success: false
                    data: null
                    error:
                      code: ACCOUNT_REJOIN_BLOCKED
                      message: 탈퇴 후 7일 동안 재가입할 수 없습니다.
  /api/v1/auth/login:
    post:
      tags:
        - Auth
      summary: 로그인
      description: 이메일과 비밀번호를 검증한 뒤 Access Token은 응답 본문으로, Refresh Token은 HttpOnly 쿠키로 발급합니다.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: test@example.com
              password: password123!
      responses:
        '200':
          description: 로그인 성공
          headers:
            Set-Cookie:
              description: HttpOnly Refresh Token 쿠키. Domain은 설정하지 않으며 Path는 `/api/v1/auth`입니다.
              schema:
                type: string
              example: gather_refresh_token=refresh-token-value; Path=/api/v1/auth; Max-Age=1209600; HttpOnly; SameSite=Lax
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenApiResponse'
              example:
                success: true
                data:
                  accessToken: access-token-value
                  tokenType: Bearer
                error: null
        '400':
          description: 요청 값 검증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
              example:
                success: false
                data: null
                error:
                  code: VALIDATION_ERROR
                  message: 요청 값이 올바르지 않습니다.
        '401':
          description: 로그인 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
              example:
                success: false
                data: null
                error:
                  code: INVALID_LOGIN
                  message: 이메일 또는 비밀번호가 올바르지 않습니다.
        '403':
          description: 계정 상태로 인한 로그인 차단
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
              examples:
                suspendedUser:
                  summary: 이용 정지 계정
                  value:
                    success: false
                    data: null
                    error:
                      code: SUSPENDED_USER
                      message: 이용 정지된 계정입니다.
                withdrawalPendingUser:
                  summary: 탈퇴 처리 중인 계정
                  value:
                    success: false
                    data: null
                    error:
                      code: WITHDRAWAL_PENDING_USER
                      message: 탈퇴 처리 중인 계정입니다.
                withdrawnUser:
                  summary: 탈퇴 계정
                  value:
                    success: false
                    data: null
                    error:
                      code: WITHDRAWN_USER
                      message: 탈퇴한 계정입니다.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Gender:
      type: string
      description: 성별
      enum:
        - MALE
        - FEMALE
      example: MALE
...
    ErrorApiResponse:
      type: object
      required:
        - success
        - data
        - error
      properties:
        success:
          type: boolean
          example: false
        data:
          type: object
          nullable: true
          example: null
        error:
          $ref: '#/components/schemas/ApiError'
...
    SignupRequest:
      type: object
      required:
        - name
        - birthDate
        - gender
        - phoneNumber
        - phoneVerificationId
        - email
        - password
        - passwordConfirm
        - nickname
        - activityRegionId
        - interestCategories
        - serviceTermsAgreed
        - privacyPolicyAgreed
        - marketingAgreed
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 20
          pattern: '^(?:[가-힣]{2,10}|[A-Za-z]{2,20})$'
          description: 이름. 완성형 한글 2~10자 또는 영문 2~20자만 허용하며, 혼합·공백·숫자·특수문자는 허용하지 않습니다.
          example: 홍길동
        birthDate:
          type: string
          format: date
          description: 생년월일. yyyy-MM-dd 형식이며 미래 날짜는 허용하지 않습니다.
          example: '2002-03-15'
        gender:
          $ref: '#/components/schemas/Gender'
        phoneNumber:
          type: string
          maxLength: 20
          description: 전화번호. 하이픈과 공백을 제거한 결과가 010으로 시작하는 11자리 숫자여야 합니다. +82는 자동 변환하지 않습니다.
          example: '01012345678'
        phoneVerificationId:
          type: string
          format: uuid
          description: 이 전화번호를 VERIFIED로 만든 인증 세션 ID. 가입 성공 트랜잭션에서 한 번만 소비됩니다.
          example: 5c5d5db1-4187-43d0-8580-672307994878
        email:
          type: string
          format: email
          maxLength: 255
          description: 이메일. 이메일 인증 완료가 필요합니다. 서버에서 앞뒤 공백 제거 및 소문자화합니다.
          example: test@example.com
        password:
          type: string
          minLength: 6
          maxLength: 12
          description: 비밀번호. 6자 이상 12자 이하입니다.
          example: password123!
        passwordConfirm:
          type: string
          minLength: 6
          maxLength: 12
          description: 비밀번호 확인
          example: password123!
        nickname:
          type: string
          minLength: 2
          maxLength: 20
          pattern: '^(?:[가-힣]{2,10}|[A-Za-z]{2,20})$'
          description: 닉네임. 완성형 한글 2~10자 또는 영문 2~20자만 허용하며, 혼합·공백·숫자·특수문자는 허용하지 않습니다.
          example: 길동
        introduction:
          type: string
          maxLength: 50
          nullable: true
          description: 소개글. 최대 50자입니다.
          example: 함께 봉사하고 싶습니다.
        activityRegionId:
          type: integer
          format: int64
          description: 활동 지역 ID. 시도(level 1) 또는 시군구(level 2) 단위 활동 지역 1개입니다. 향후 공고/모임 검색의 기본 지역 필터 초기값으로 사용합니다.
          example: 123
        interestCategories:
          type: array
          minItems: 1
          uniqueItems: true
          description: 관심 카테고리 목록. 중복 없이 1개 이상입니다.
          items:
            $ref: '#/components/schemas/PostingCategory'
          example:
            - WELFARE
            - EDUCATION
        serviceTermsAgreed:
          type: boolean
          description: 서비스 이용약관 동의 여부. true 필수입니다.
          example: true
        privacyPolicyAgreed:
          type: boolean
          description: 개인정보 처리방침 동의 여부. true 필수입니다.
          example: true
        marketingAgreed:
          type: boolean
          description: 맞춤형 봉사/이벤트 알림 수신 동의 여부. 선택 동의입니다.
          example: false
...
    SignupApiResponse:
      type: object
      required:
        - success
        - data
        - error
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/SignupData'
        error:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ApiError'
          example: null
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          description: 가입한 이메일
          example: test@example.com
        password:
          type: string
          description: 비밀번호
          example: password123!
...